CRUD productos (solo crear) y detalles
Se implemento el formulario para crear un producto mediante la dashboard de administrador y todo lo necesario para esto (métodos del controlador, vistas, modelos, etc)

+++ database/migrations/2024_05_16_162413_create_beerformats_table.php
@@ -16,7 +16,7 @@ class CreateBeerformatsTable extends Migration
         Schema::create('beerformats', function (Blueprint $table) {
             $table->id();
             $table->string('container');
-            $table->integer('liters');
+            $table->decimal('liters', 5, 3);
             $table->timestamps();
         });
     }
